@@ -1,6 +1,6 @@
 module github.com/acouvreur/sablier/plugins/caddy
 
-go 1.18
+go 1.21
 
 require github.com/caddyserver/caddy/v2 v2.6.4
 
