@@ -1,3 +1,3 @@
 module github.com/acouvreur/sablier/plugins/traefik
 
-go 1.18
+go 1.21
